summarizer_agent/requirements.txt
google-adk>=1.28.0,<2.0.0
